add auth to delete command handler

+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -1,14 +1,17 @@
 using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Core.Domain.Constants;
 using Restaurants.Core.Domain.Entities;
 using Restaurants.Core.Domain.RepositoryContracts;
+using Restaurants.Core.Domain.ServiceContracts;
 using Restaurants.Core.Exceptions;
 
 namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant;
 
 public class DeleteRestaurantCommandHandler(
     ILogger<DeleteRestaurantCommandHandler> logger,
-    IRestaurantsRepository repository
+    IRestaurantsRepository repository,
+    IRestaurantAuthorizationService authorizationService
 ) : IRequestHandler<DeleteRestaurantCommand>
 {
     public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
@@ -19,6 +22,8 @@ public class DeleteRestaurantCommandHandler(
 
         if (restaurant is null) throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
 
+        if (authorizationService.Authorize(restaurant, ResourceOperation.Delete)) throw new ForbidException();
+        
         await repository.DeleteAsync(restaurant);
     }
 }
